Update Series.reverse() to use apply_gather
Test cases:
  test_series_nlargest
  test_series_nsmallest

`cudautils.reverse_array` never used any more

+++ python/cudf/utils/cudautils.py
@@ -57,7 +57,7 @@ def gpu_arange_reversed(size, out):
         out[i] = size - i - 1
 
 
-def arange_reversed(size, dtype=np.int64):
+def arange_reversed(size, dtype=np.int32):
     out = rmm.device_array(size, dtype=dtype)
     if size > 0:
         gpu_arange_reversed.forall(size)(size, out)
